Drop schema and related tables in tests migration
- This also fixes an issue with the previous migration that wasn't
apparent. All the migrations were being added to the same sequence due
to the use of a variable to store a reference to Bookshelf.knex.schema
which wasn't producing a reference at all but initiating a new
SchemaBuilder object separate from Bookshelf.knex.schema which wasn't
being cleared each time the sequence ran. This meant that all the
queries were being piled up and ran multiples times instead of the
expected single run.

This wasn't causing any major issues because of the dropTableIfExists
method which doesn't throw if the tables don't exist. However, once
schemas got added to the mix this meant that the migration tried to
also drop and create all tables from the 'public' schema in the 'test'
schema as well, which was unexpected and would eventually produce an
error.